move each component from Popup.js file to separate files
