+++ imports/ui/pages/components/Popup/index.js
@@ -1,0 +1,13 @@
+import GameResignedPopup from "./GameResignedPopup";
+import GameRequestPopup from "./GameRequestPopup";
+import ExaminActionPopup from "./ExaminActionPopup";
+import ActionPopup from "./ActionPopup";
+import GameNotificationPopup from "./GameNotificationPopup";
+
+export {
+  GameResignedPopup,
+  ExaminActionPopup,
+  GameRequestPopup,
+  ActionPopup,
+  GameNotificationPopup
+};
